move shared components to shared_component package
