+++ feature/media_details/presentation/src/main/java/com/sanaa/presentation/screen/episode_details/EpisodeDetailsScreen.kt
@@ -28,12 +28,12 @@ import com.sanaa.designsystem.design_system.component.novix_scaffold.NovixScaffo
 import com.sanaa.designsystem.design_system.component.top_bar.NovixTopBar
 import com.sanaa.designsystem.design_system.component.top_bar.TopBarClickableIcon
 import com.sanaa.feature.mediadetails.presentation.R
-import com.sanaa.presentation.component.OverviewSection
-import com.sanaa.presentation.component.RequestToLoginBottomSheet
+import com.sanaa.presentation.shared_component.OverviewSection
+import com.sanaa.presentation.shared_component.RequestToLoginBottomSheet
 import com.sanaa.presentation.navigation.ActorDetailsScreenRoute
 import com.sanaa.presentation.navigation.LocalNavControllerProvider
 import com.sanaa.presentation.screen.episode_details.components.GuestsOfHonorComponent
-import com.sanaa.presentation.screen.series.components.BottomContainer
+import com.sanaa.presentation.shared_component.BottomContainer
 import com.sanaa.presentation.screen.series.components.SeriesHeaderSection
 import kotlinx.coroutines.flow.collectLatest
 import org.koin.androidx.compose.koinViewModel
