+++ feature/media_details/presentation/src/main/java/com/sanaa/presentation/screen/series/components/CastComponent.kt
@@ -15,7 +15,7 @@ import androidx.compose.ui.unit.dp
 import coil.compose.rememberAsyncImagePainter
 import com.sanaa.designsystem.design_system.theme.Theme
 import com.sanaa.feature.mediadetails.presentation.R
-import com.sanaa.presentation.component.cards.ActorCard
+import com.sanaa.presentation.shared_component.cards.ActorCard
 import com.sanaa.presentation.model.ActorUiModel
 
 @Composable
