+++ feature/search/presentation/src/main/java/com/sanaa/presentation/screen/componants/SearchSection.kt
@@ -45,7 +45,7 @@ fun SearchSection(
     Row(
         modifier = Modifier
             .height(48.dp)
-            .padding(start = 16.dp, end = 16.dp, bottom = 8.dp)
+            .padding(horizontal = 16.dp)
             .fillMaxWidth(),
         horizontalArrangement = Arrangement.spacedBy(space = 8.dp),
         verticalAlignment = Alignment.CenterVertically
